Add end to end test
This new test mocks both the user agent (e.g. browser) and the push service
(e.g. Firestore) to verify that encryption and decryption works properly.

I used the RFCs as reference (RFC8291, RFC8292 & RFC 8188), but didn't follow
them to the letter. The result can successfully check all the signatures and
decrypt the content, so it seems to be working.

Instead of the deprecated crypto/elliptic functions, this makes heavy use of
crypto/ecdh, which require Go 1.20. But this is only a test dependency, library
users should not be impacted.